+++ src/services/router.js
@@ -1,6 +1,6 @@
 import Vue from 'vue'
 import VueRouter from 'vue-router'
-import createRoutes from '../routes'
+import createRoutes from '../routes/index'
 
 Vue.use(VueRouter)
